move multi and child sensors to a folder
