+++ src/Win2Mqtt.Common/SystemSensors/Multi/SystemMultiSensor.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+using System;
+
+namespace Win2Mqtt.SystemSensors.Multi
+{
+    public abstract class SystemMultiSensor : ISystemMultiSensor
+    {
+        protected SystemMultiSensor()
+        {
+            Key = SensorMetadataFactory.GetKeyFromAttribute(this)!;
+        }
+
+        public string Key { get; set; }
+
+        public abstract IEnumerable<ISystemSensor> CreateSensors(IServiceProvider serviceProvider);
+    }
+}
