+++ src/Win2Mqtt.Common/SystemSensors/Multi/SystemMultiSensorAttribute.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Win2Mqtt.SystemSensors.Multi
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public sealed class SystemMultiSensorAttribute(string key) : Attribute
+    {
+        public string Key { get; } = key;
+    }
+
+}
