+++ src/Win2Mqtt.Common/SystemSensors/SensorMetadataFactory.cs
@@ -8,7 +8,7 @@ namespace Win2Mqtt.SystemSensors
         public static SystemSensorMetadata? FromSensor(ISystemSensor sensor)
         {
             var type = sensor.GetType();
-            var attr = type.GetCustomAttribute<SensorAttribute>();
+            var attr = type.GetCustomAttribute<SystemSensorAttribute>();
             if (attr == null) return null;
 
             var valueType = type.GetInterfaces()
@@ -26,34 +26,5 @@ namespace Win2Mqtt.SystemSensors
                 IsBinary = attr.IsBinary
             };
         }
-        public static SystemSensorMetadata? FromChildSensor(ISystemSensor sensor, string id)
-        {
-            var type = sensor.GetType();
-            var attr = type.GetCustomAttribute<ChildSensorAttribute>();
-            if (attr == null) return null;
-
-            var valueType = type.GetInterfaces()
-                .FirstOrDefault(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(ISystemSensor<>))
-                ?.GenericTypeArguments[0] ?? typeof(object);
-
-            return new SystemSensorMetadata
-            {
-                Key = string.Format(attr.KeyPattern, id),
-                Name = attr.NamePattern == null ? string.Format(attr.KeyPattern, id) : string.Format(attr.NamePattern, id),
-                ValueType = valueType,
-                UnitOfMeasurement = attr.UnitOfMeasurement,
-                DeviceClass = attr.DeviceClass,
-                StateClass = attr.StateClass,
-                IsBinary = attr.IsBinary
-            };
         }
-
-        public static string? GetKeyFromAttribute(ISystemMultiSensor multiSensor)
-        {
-            var type = multiSensor.GetType();
-            var attr = type.GetCustomAttribute<MultiSensorAttribute>();
-            if (attr == null) return null;
-            return attr.Key;
-        }
-    }
 }
