+++ src/Win2Mqtt.SystemSensors.Windows/MultiSensors/Drive/DriveFreeSizeSensor.cs
@@ -1,14 +1,15 @@
 ﻿using Microsoft.Extensions.Logging;
+using Win2Mqtt.SystemSensors.Multi;
 
 namespace Win2Mqtt.SystemSensors.Windows.MultiSensors.Drive
 {
-    [ChildSensor("drive/{0}/sizefree",
+    [SystemChildSensor("drive/{0}/sizefree",
     namePattern: "Drive {0} Free Space",
     unitOfMeasurement: "B",
     deviceClass: "data_size",
     stateClass: "measurement")]
     public class DriveFreeSizeSensor(DriveInfo driveInfo, ILogger<DriveFreeSizeSensor> logger) 
-        : ChildSensor<long>(driveInfo.Name.Replace(":\\", ""))
+        : SystemChildSensor<long>(driveInfo.Name.Replace(":\\", ""))
     {
         public override Task<long> CollectAsync()
         {
