+++ src/Win2Mqtt.SystemSensors.Windows/MultiSensors/Drive/DriveTotalSizeSensor.cs
@@ -1,14 +1,15 @@
 ﻿using Microsoft.Extensions.Logging;
+using Win2Mqtt.SystemSensors.Multi;
 
 namespace Win2Mqtt.SystemSensors.Windows.MultiSensors.Drive
 {
-    [ChildSensor("drive/{0}/sizetotal",
+    [SystemChildSensor("drive/{0}/sizetotal",
     namePattern: "Drive {0} Total Space",
     unitOfMeasurement: "B",
     deviceClass: "data_size",
     stateClass: "measurement")]
     public class DriveTotalSizeSensor(DriveInfo driveInfo, ILogger<DriveTotalSizeSensor> logger) 
-        : ChildSensor<long>(driveInfo.Name.Replace(":\\", ""))
+        : SystemChildSensor<long>(driveInfo.Name.Replace(":\\", ""))
     {
         public override Task<long> CollectAsync()
         {
