+++ src/Win2Mqtt.SystemSensors.Windows/Sensors/FreeMemorySensor.cs
@@ -3,8 +3,8 @@ using System.Runtime.InteropServices;
 
 namespace Win2Mqtt.SystemSensors.Windows.Sensors
 {
-    [Sensor("freememory", name: "Free Memory", unitOfMeasurement: "B", deviceClass: "memory", stateClass: "measurement")]
-    public partial class FreeMemorySensor(ILogger<FreeMemorySensor> logger) : Sensor<double>
+    [SystemSensor("freememory", name: "Free Memory", unitOfMeasurement: "B", deviceClass: "memory", stateClass: "measurement")]
+    public partial class FreeMemorySensor(ILogger<FreeMemorySensor> logger) : SystemSensor<double>
     {
         public override async Task<double> CollectAsync()
         {
